refactor(net): move `RecvFrom` trait to `ipv6` module
